% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_colnames.R
\name{check_colnames}
\alias{check_colnames}
\title{Check colnames of imported file}
\usage{
check_colnames(filename, sheetname, skiprows, oldnames, newnames)
}
\arguments{
\item{filename}{Filename (and path) of dataset to import}

\item{sheetname}{Sheetname}

\item{skiprows}{Number of rows to skip}

\item{oldnames}{Variable names to change}

\item{newnames}{New variable names}
}
\value{
Message if colnames of imported file match 'oldnames' and a dataframe
containing the colnames of the imported file, 'oldnames' and 'newnames'.
}
\description{
Check colnames of imported file
}
