% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop_sensitive.R
\name{drop_sensitive}
\alias{drop_sensitive}
\title{Drop sensitive data}
\usage{
drop_sensitive(
  data,
  drop = c("ahvnr", "firstname", "surname", "birthday"),
  drop_add = NA
)
}
\arguments{
\item{data}{Dataframe}

\item{drop}{Variables to drop. Default: c("ahvnr", "firstname", "surname", "birthday")}

\item{drop_add}{Additional variables to drop}
}
\value{
Dataset from which indicated variables have been dropped.
}
\description{
Drop sensitive data
}
