% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_pseudonymized.R
\name{export_pseudonymized}
\alias{export_pseudonymized}
\title{Export pseudonymized data to csv}
\usage{
export_pseudonymized(data, sensitive_vars, path, data_name, data_summary)
}
\arguments{
\item{data}{Dataframe}

\item{sensitive_vars}{Sensitive variables to drop. These are not exported.}

\item{path}{File path to directory where data will be stored. The function
will create a subfolder folder called 'panon' and store the table there (i.e.
in 'path/panon').}

\item{data_name}{Name or source of data}

\item{data_summary}{TRUE (Default) to print head of exported data.}
}
\value{
Exported csv-file from which indicated variables have been dropped.
}
\description{
Export pseudonymized data to csv
}
\examples{
\dontrun{
df <- unique_id(data = testdata, id = "ahvnr", salt = "myverygoodsalt")
df <- export_pseudonymized(data = df,
                           sensitive_vars = c("ahvnr","firstname","surname",
                                              "birthday","plz", "wohnort"),
                           path = "output",
                           data_name = "FAKE_DATA",
                           data_summary = TRUE)
}
}
