% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudonymize.R
\name{pseudonymize}
\alias{pseudonymize}
\title{Pseudonymize data}
\usage{
pseudonymize(
  data_name,
  import_filename,
  import_sheetname = NULL,
  import_skiprows = 0,
  import_oldnames,
  import_newnames,
  id_original,
  id_salt,
  id_expected_format = "digit",
  id_expected_length = 13,
  date_var = "birthday",
  date_new_format = "\%Y",
  date_new_var = "birthyear",
  export_path,
  export_address = TRUE,
  export_address_vars,
  sensitive,
  data_summary = TRUE
)
}
\arguments{
\item{data_name}{Name of data}

\item{import_filename}{Filename (and path) of dataset to import}

\item{import_sheetname}{Sheetname}

\item{import_skiprows}{Number of rows to skip}

\item{import_oldnames}{Variable names to change}

\item{import_newnames}{New variable names}

\item{id_original}{ID-variable that needs to by pseudonymized}

\item{id_salt}{Salt value}

\item{id_expected_format}{Expected format of ID-variable (Default: 'digit').
If the ID-variable does not have the expected format, the function will try to
convert it.}

\item{id_expected_length}{Expected length of ID-variable (Default: 13). If the
ID-variable does not have the expected length, the function will give an error.}

\item{date_var}{Date variable}

\item{date_new_format}{Format for new date variable}

\item{date_new_var}{Name for new date variable}

\item{export_path}{Path exported data}

\item{export_address}{TRUE (Default) to export address variables}

\item{export_address_vars}{Variables to export for address file}

\item{sensitive}{Variables to drop.}

\item{data_summary}{TRUE (Default) to print head of pseudonymized data.}
}
\value{
Export of following files: keytable, address file, pseudonymized data
}
\description{
Pseudonymize data
}
