% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{testdata}
\alias{testdata}
\title{Testdata for the package}
\format{
An object of class \code{data.frame} with 3 rows and 14 columns.
}
\usage{
testdata
}
\description{
This is some fake dataset that is similar to the real data.
}
\keyword{datasets}
