% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unique_geo_id.R
\name{unique_geo_id}
\alias{unique_geo_id}
\title{Create unique IDs for address and house}
\usage{
unique_geo_id(
  df,
  var_street = "street",
  var_number = "street_nr",
  var_plz = "plz",
  list_rename = c("egid", "ewid"),
  list_distinct = c(999999999, 999)
)
}
\arguments{
\item{df}{Dataframe}

\item{var_street}{Street}

\item{var_number}{Street number}

\item{var_plz}{PLZ}

\item{var_egid}{EGID}

\item{var_ewid}{EWID}
}
\value{
Dataframe with additional IDs
}
\description{
These IDs are only unique within a dataset, they are not consistent
throughout different datasets or years.
}
