% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unique_id.R
\name{unique_id}
\alias{unique_id}
\title{Generate unique ID from AHV-number}
\usage{
unique_id(
  data,
  id = "ahvnr",
  salt,
  id_expected_format = "digit",
  id_expected_length = 13
)
}
\arguments{
\item{data}{Dataframe}

\item{id}{ID-variable that needs to by pseudonymized}

\item{salt}{Salt value}

\item{id_expected_format}{Expected format of ID-variable (Default: 'digit').
If the ID-variable does not have the expected format, the function will try to
convert it.}

\item{id_expected_length}{Expected length of ID-variable (Default: 13). If the
ID-variable does not have the expected length, the function will give an error.}
}
\value{
Dataframe with pseudo_id as additional variable. If id does not have
the expected format, the dataframe will contain the converted value.
}
\description{
Generate unique ID from AHV-number
}
