% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unique_geo_id.R
\name{unique_id_simple}
\alias{unique_id_simple}
\title{Create a simple, numeric id}
\usage{
unique_id_simple(df, cols)
}
\arguments{
\item{df}{Dataframe}

\item{cols}{Variables to use for unique id}
}
\value{
Vector with the new id
}
\description{
Create a simple, numeric id
}
