% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_sensitive.R
\name{aggregate_sensitive}
\alias{aggregate_sensitive}
\title{Aggregate sensitive data}
\usage{
aggregate_sensitive(
  data,
  date_var = "birthday",
  date_old_format = "\%Y-\%m-\%d",
  date_new_format = "\%Y",
  date_new_var = "birthyear",
  var_zip = "plz",
  var_new_zip = "plz4"
)
}
\arguments{
\item{data}{data frame}

\item{date_var}{Date variable}

\item{date_old_format}{Format of old date variable}

\item{date_new_format}{Format for new date variable}

\item{date_new_var}{Name for new date variable}

\item{var_zip}{Variable for zip code}
}
\value{
datafile for which indicated variables have been aggregated
}
\description{
Aggregate sensitive data
}
