% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_sensitive.R
\name{aggregate_sensitive}
\alias{aggregate_sensitive}
\title{Aggregate sensitive data}
\usage{
aggregate_sensitive(
  data,
  date_var = "birthday",
  date_old_format = "\%d.\%m.\%Y",
  date_new_format = "\%Y",
  date_new_var = "birthyear"
)
}
\arguments{
\item{data}{Dataframe}

\item{date_var}{Date variable(s)}

\item{date_old_format}{Format of old date variable(s)}

\item{date_new_format}{Format for new date variable(s)}

\item{date_new_var}{Name for new date variable(s)}
}
\value{
Original dataframe plus the new date variable(s)
}
\description{
This function aggregates a date variable (e.g. date of birth to year of birth).
}
\examples{
testdata <- cbind(testdata,
                  data.frame("other_date" = c("20.03.2000", "12.06.1998", "26.12.1989")))
aggregate_sensitive(data = testdata,
                    date_var = c("birthday", "other_date"),
                    date_old_format = c("\%Y-\%m-\%d", "\%d.\%m.\%Y"),
                    date_new_format = c("\%Y", "\%Y"),
                    date_new_var = c("birthyear", "otheryear"))
}
