% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_sensitive.R
\name{aggregate_sensitive}
\alias{aggregate_sensitive}
\title{Aggregate sensitive data}
\usage{
aggregate_sensitive(
  data,
  date_var = "birthday",
  date_new_format = "\%Y",
  date_new_var = "birthyear"
)
}
\arguments{
\item{data}{Dataframe}

\item{date_var}{Date variable}

\item{date_new_format}{Format for new date variable}

\item{date_new_var}{Name for new date variable}
}
\value{
Datafile for which indicated variables have been aggregated
}
\description{
Aggregate sensitive data
}
