% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/append_keytable.R
\name{append_keytable}
\alias{append_keytable}
\title{Generate and append key table}
\usage{
append_keytable(df, path)
}
\arguments{
\item{df}{Dataframe}

\item{path}{Filepath}
}
\value{
Writes or appends keytable
}
\description{
Generate and append key table
}
