% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_address.R
\name{export_address}
\alias{export_address}
\title{Generate address file}
\usage{
export_address(data, path, data_name, vars)
}
\arguments{
\item{data}{Dataframe}

\item{path}{Filepath}

\item{data_name}{Name or source of data}

\item{vars}{Variables to export}
}
\value{
Exported csv-file.
}
\description{
Generate address file
}
