% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_address.R
\name{export_address}
\alias{export_address}
\title{Generate address file}
\usage{
export_address(data, filename, vars)
}
\arguments{
\item{data}{Dataframe}

\item{filename}{Filename (and path) for exported address file}

\item{vars}{Variables to export}
}
\value{
Exported csv-file.
}
\description{
Generate address file
}
