% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_pseudonymized.R
\name{export_pseudonymized}
\alias{export_pseudonymized}
\title{Export pseudonymized data to csv}
\usage{
export_pseudonymized(
  data,
  sensitive = c("ahvnr", "firstname", "surname", "birthday"),
  path,
  data_name,
  data_summary
)
}
\arguments{
\item{data}{Dataframe}

\item{sensitive}{Sensitive variables to drop. Default: c("ahvnr", "firstname", "surname", "birthday")}

\item{path}{Filepath}

\item{data_name}{Name or source of data}

\item{data_summary}{TRUE (Default) to print head of exported data.}
}
\value{
Exported csv-file from which indicated variables have been dropped.
}
\description{
Export pseudonymized data to csv
}
