% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_raw_data.R
\name{import_raw_data}
\alias{import_raw_data}
\title{Import datasets into R}
\usage{
import_raw_data(filename, sheetname = NULL, skiprows = 0, oldnames, newnames)
}
\arguments{
\item{filename}{Filename (and path) of dataset to import}

\item{sheetname}{Sheetname}

\item{skiprows}{Number of rows to skip}

\item{oldnames}{Variable names to change}

\item{newnames}{New variable names}
}
\value{
Imported dataset, with renamed variables.
}
\description{
Import datasets into R
}
