% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudonymize.R
\name{pseudonymize}
\alias{pseudonymize}
\title{Pseudonymize data}
\usage{
pseudonymize(
  data_name,
  import_filename,
  import_sheetname = NULL,
  import_skiprows = 0,
  import_oldnames,
  import_newnames,
  id,
  id_salt,
  id_expected_format = "digit",
  id_expected_length = 13,
  date_var = "birthday",
  date_new_format = "\%Y",
  date_new_var = "birthyear",
  var_street = "strasse",
  var_number = "hausnr",
  var_plz = "plz",
  var_egid = "egid",
  var_ewid = "ewid",
  zip_var = "plz",
  zip_new_var = "plz4",
  path,
  export_address = TRUE,
  address_vars,
  keytable_vars,
  sensitive_vars,
  data_summary = TRUE
)
}
\arguments{
\item{data_name}{Name of data}

\item{import_filename}{Filename (and path) of dataset to import}

\item{import_sheetname}{Sheetname}

\item{import_skiprows}{Number of rows to skip}

\item{import_oldnames}{Variable names to change}

\item{import_newnames}{New variable names}

\item{id}{ID-variable that needs to by pseudonymized}

\item{id_salt}{Salt value}

\item{id_expected_format}{Expected format of ID-variable (Default: 'digit').
If the ID-variable does not have the expected format, the function will try to
convert it.}

\item{id_expected_length}{Expected length of ID-variable (Default: 13). If the
ID-variable does not have the expected length, the function will give an error.}

\item{date_var}{Date variable}

\item{date_new_format}{Format for new date variable}

\item{date_new_var}{Name for new date variable}

\item{var_street}{Variable for street name}

\item{var_number}{Variable for street number}

\item{var_plz}{Variable for zip code}

\item{var_egid}{Variable of official building registry}

\item{var_ewid}{Variable of official flat registry}

\item{path}{Path exported data}

\item{export_address}{TRUE (Default) to export address variables}

\item{address_vars}{Variables to export for address file}

\item{keytable_vars}{Variables to export for keytable}

\item{sensitive_vars}{Variables to drop from final pseudonymized data.}

\item{data_summary}{TRUE (Default) to print head of pseudonymized data.}
}
\value{
Export of following files: keytable, address file, pseudonymized data
}
\description{
Pseudonymize data
}
