% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unique_id.R
\name{unique_id}
\alias{unique_id}
\title{Generate unique, pseudonymized ID}
\usage{
unique_id(data, id = "ahvnr", salt, id_expected_format = "digit")
}
\arguments{
\item{data}{Dataframe}

\item{id}{ID-variable that needs to by pseudonymized}

\item{salt}{Salt value}

\item{id_expected_format}{Expected pattern of ID-variable strings. If the ID-variable
does not have the expected pattern, the function will remove any characters that are
not of the expected format (Default = 'digit', i.e. letters, punctuation and spaces
will be removed from the original ID before applying the hash function).}
}
\value{
Dataframe with pseudo_id as additional variable. If id does not have
the expected format, the dataframe will contain the corrected value.
}
\description{
This function generates a unique, pseudonymized identifier using SHA-256,
a vectorized hash function. Choose a random key ('salt') that is only known to
you for additional security.
}
\examples{
unique_id(data = testdata, id = "ahvnr", salt = "myverygoodsalt")
}
